+++ src/main/java/tanks/network/event/EventTankUpdateHealth.java
@@ -37,7 +37,7 @@ public class EventTankUpdateHealth extends PersonalEvent
 		double before = t.health;
 		t.health = health;
 
-		if ((int) (before) != (int) (t.health) && t.health > 0)
+		if ((int) (before) != (int) (t.health) && (int) t.health > 0)
 		{
 			Effect e = Effect.createNewEffect(t.posX, t.posY, t.posZ + t.size * 0.75, Effect.EffectType.shield);
 			e.size = t.size;
